requirements.txt
mcp
uvicorn
fastapi
python-dotenv
openai
azure-search-documents
azure-core
azure-identity
pandas
openpyxl